Moved everything from the `f32` module up.
