--- src/f32/mod.rs
@@ -1,27 +1,0 @@
-mod cast;
-#[cfg(feature = "transform-types")]
-mod transform;
-
-pub use cast::{F32x16Cast, F32x2Cast, F32x3Cast, F32x4Cast, F32x9Cast};
-#[cfg(feature = "transform-types")]
-pub use transform::*;
-
-#[cfg(feature = "bytemuck")]
-mod glam_bytemuck;
-#[cfg(feature = "bytemuck")]
-pub use glam_bytemuck::*;
-
-#[cfg(feature = "mint")]
-mod glam_mint;
-#[cfg(feature = "mint")]
-pub use glam_mint::*;
-
-#[cfg(feature = "rand")]
-mod glam_rand;
-#[cfg(feature = "rand")]
-pub use glam_rand::*;
-
-#[cfg(feature = "serde")]
-mod glam_serde;
-#[cfg(feature = "serde")]
-pub use glam_serde::*;
